fix(npmPublish): remove npm publish from release step
the previous attempt to release was done under CI where npm token was deleted. The build failed
because no token specified. In order for the build to pass, a rule to not run the npm publish script
needs to be used. This is an attempt to do it from package json

BREAKING CHANGE: npm publish